backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
tavily-python
python-dotenv
PyPDF2
python-docx
requests
reportlab
fastapi-mail
